Fixed broken "PUBLICATIONS on DBLP" link.
This is [the old "PUBLICATIONS on DBLP" link address][old_link] on the
[db.cs.berkeley.edu][dbsite] website:

```
http://dblp.org/search/index.php#query=author:michael_j_franklin|author:joseph_m_hellerstein|author:ion_stoica|author:joseph_e_gonzalez|author:ali_ghodsi|author:eric_brewer|author:raluca_ada_popa|author:rachit_agarwal|author:_|author:_|author:julian_shun|author:sara_alspaugh|author:daniel_crankshaw|author:daniel_haas|author:sanjay_krishnan|author:nick_lanham|author:evan_sparks|author:liwen_sun|author:shivaram_venkataraman|author:johann_schleier_smith|author:_|author:_|author:_|author:vikram_sreekanti|author:michael_whittaker|author:larry_xu|author:peter_alvaro|author:sameer_agarwal|author:peter_bailis|author:daniel_bruckner|author:neil_conway|author:tim_kraska|author:gene_pang|author:josh_rosen|author:ameet_talkwalkar|author:beth_trushkowsky|author:reynold_xin|author:christos_kozanitis|author:jiannan_wang|author:_|author:tomer_kaftan|
```

Clicking on the link takes you to dblp with the search field filled in, but you
are met with a sobering *no matches* result.

First, I went to [Michael Franklin's dblp page][franklin]. Next to the "Michael
J. Franklin" text at the top is a "little magnifying glass with rightward
arrow" icon. Mousing over that icon produced a drop-down menu with the last
entry being a link to [view the author in dblp search][franklin_search].
Clicking on this link brings you to the first of the following two equivalent
URLs:

```
http://dblp.dagstuhl.de/search/publ?q=author:Michael_J._Franklin:
http://dblp.org/search/publ?q=author:Michael_J._Franklin:
```

The following URLs bring you somewhere else:

- `http://dblp.org/search/publ?q=author:Michael_J._Franklin` (missing second
  colon) returns more search results.
- `http://dblp.org/search/publ?q=author:michael_J._Franklin:` (lowercase m)
  returns *no matches*
- `http://dblp.org/search/publ?q=author:Michael_J_Franklin:` (missing period
  after J) returns *no matches*

From this tinkering, it seems like dblp is somewhat particular about
capitalization and punctuation. Let's take the old link, add colons to the end
of each author, change the `index.php#query=` bit to `publ?q=`, and correct the
capitalization and punctuation of Michael Franklin. Here's [what we
get][http://bit.ly/2lBUinh]:

```
http://dblp.org/search/publ?q=author:Michael_J._Franklin:|author:joseph_m_hellerstein:|author:ion_stoica:|author:joseph_e_gonzalez:|author:ali_ghodsi:|author:eric_brewer:|author:raluca_ada_popa:|author:rachit_agarwal:|author:_:|author:_:|author:julian_shun:|author:sara_alspaugh:|author:daniel_crankshaw:|author:daniel_haas:|author:sanjay_krishnan:|author:nick_lanham:|author:evan_sparks:|author:liwen_sun:|author:shivaram_venkataraman:|author:johann_schleier_smith:|author:_:|author:_:|author:_:|author:vikram_sreekanti:|author:michael_whittaker:|author:larry_xu:|author:peter_alvaro:|author:sameer_agarwal:|author:peter_bailis:|author:daniel_bruckner:|author:neil_conway:|author:tim_kraska:|author:gene_pang:|author:josh_rosen:|author:ameet_talkwalkar:|author:beth_trushkowsky:|author:reynold_xin:|author:christos_kozanitis:|author:jiannan_wang:|author:_:|author:tomer_kaftan:
```

Follow this link and you'll see we get results for Michael Franklin!

This PR changes `index.html` and `_data/people.yml` to make sure the produced
dblp link is in the correct form.

[old_link]:        http://bit.ly/2kfJGya
[dbsite]:          http://db.cs.berkeley.edu/
[franklin]:        http://dblp.dagstuhl.de/pers/hd/f/Franklin:Michael_J=
[franklin_search]: http://dblp.dagstuhl.de/search/publ?q=author:Michael_J._Franklin:
[franklin_fixed]:  http://bit.ly/2lBUinh